Floating point release

+++ sbs_deployment_framework/sbs_sw_stack/libs/sbs_neural_network/inc/sbs_neural_network.h
@@ -33,7 +33,7 @@ typedef enum
   COLUMN_SHIFT
 } WeightShift;
 
-typedef uint32_t  	NeuronState;
+typedef float     	NeuronState;
 typedef uint32_t    Epsilon;
 typedef void * 		  SbsWeightMatrix;
